Added chrome and safari base classes
